semantic-search-service/requirements.txt
fastapi
uvicorn
python-docx
requests
sentence-transformers
faiss-cpu
numpy